Render a table to give an indication of the values to choose for the prevalence level and the top filtering percentage

Render a table to give an indication of the values to choose for the prevalence level and the top filtering percentage

Usage

choose_params_values(
  data_with_annotation,
  object_of_interest,
  sample_size,
  prev_list = c(0.2, 0.3, 0.4),
  filtering_list = c(10, 20, 30),
  graph_file = NULL,
  col_module_id,
  annotation_level
)

Arguments

data_with_annotation

Dataframe. The abundance table merged with the module names. Required format: modules are the rows and samples are the columns. The first column must be the modules name (e.g. species), the second is the module ID (e.g. msp), and each subsequent column is a sample

object_of_interest

String. The name of the bacteria or species of interest or a key word in the functional module definition

sample_size

Numeric. Number of samples in each dataset.

prev_list

List of numeric. The prevalences to be studied. Required format is decimal: 0.20 for 20% of prevalence

filtering_list

List of numeric. The filtering top percentages to be studied. Required format is: 10 for the top 10%

graph_file

Dataframe. The object generated by graph_step() function

col_module_id

String. The name of the column with the module names in annotation_table

annotation_level

String. The name of the column with the level to be studied. Examples: species, genus, level_1

Value

Dataframe. Returns F1 rates before and after using NeighborFinder()

Examples

data(data)
data(graphs)
choose_params_values(data_with_annotation = data$CRC_JPN, object_of_interest = "Escherichia coli", sample_size = 100, prev_list = c(0.20, 0.30), filtering_list = c(10, 20), graph_file = graphs$CRC_JPN, col_module_id = "msp_id", annotation_level = "species")
#> Defining and saving true neighbors...
#> Calculating scores...
#>   prev_level filtering_top F1_before F1_after
#> 1        0.2            10    0.0120     0.00
#> 2        0.2            20    0.0120     0.67
#> 3        0.3            10    0.0058     0.67
#> 4        0.3            20    0.0058     1.00